Squashed 'src/monetdblite/' changes from 7329dece17..efb236f70c
efb236f70c zapping zu again
abdb644e08 Merge branch 'branches/Mar2018' of hg::https://dev.monetdb.org/hg/MonetDB
0ea7145326 Allow setting of SQL_ATTR_MAX_LENGTH and SQL_ATTR_MAX_ROW. (but only to 0 and 2147483647 (2**31 - 1).) This is one of the errors reported in the log in bug 6540.
7150846323 Always return 0 for SQL_ATTR_QUERY_TIMEOUT (no timeout). This is one of the errors reported in the log in bug 6540.
ede3ad3435 We always fill in the IPD after a SQLPrepare.
b2634804c8 Add a check that the ESCAPE string is in fact just a character.
1aa30b0251 If ESCAPE character does not occur in LIKE pattern, treat as if no ESCAPE. This fixes bug 6593.
77aee02a33 fix bug 6610

git-subtree-dir: src/monetdblite
git-subtree-split: efb236f70cd764afd04a770efcbf1c59fbb52f6e

+++ src/common/stream.c
@@ -338,8 +338,8 @@ mnstr_read(stream *restrict s, void *restrict buf, size_t elmsize, size_t cnt)
 	if (s == NULL || buf == NULL)
 		return -1;
 #ifdef STREAM_DEBUG
-	fprintf(stderr, "read %s %zu %zu\n",
-		s->name ? s->name : "<unnamed>", elmsize, cnt);
+	fprintf(stderr, "read %s %"PRIu64" %"PRIu64"\n",
+		s->name ? s->name : "<unnamed>", (uint64_t) elmsize, (uint64_t) cnt);
 #endif
 	assert(s->access == ST_READ);
 	if (s->errnr)
@@ -358,8 +358,8 @@ mnstr_readline(stream *restrict s, void *restrict buf, size_t maxcnt)
 	if (s == NULL || buf == NULL)
 		return -1;
 #ifdef STREAM_DEBUG
-	fprintf(stderr, "readline %s %zu\n",
-		s->name ? s->name : "<unnamed>", maxcnt);
+	fprintf(stderr, "readline %s %"PRIu64"\n",
+		s->name ? s->name : "<unnamed>", (uint64_t) maxcnt);
 #endif
 	assert(s->access == ST_READ);
 	if (s->errnr)
@@ -416,8 +416,8 @@ mnstr_write(stream *restrict s, const void *restrict buf, size_t elmsize, size_t
 	if (s == NULL || buf == NULL)
 		return -1;
 #ifdef STREAM_DEBUG
-	fprintf(stderr, "write %s %zu %zu\n",
-		s->name ? s->name : "<unnamed>", elmsize, cnt);
+	fprintf(stderr, "write %s %"PRIu64" %"PRIu64"\n",
+		s->name ? s->name : "<unnamed>", (uint64_t) elmsize, (uint64_t) cnt);
 #endif
 	assert(s->access == ST_WRITE);
 	if (s->errnr)
